Create ADD-Assignment3-Number1.cpp
Consider you are conducting an experiment with a single slit diffraction experiment apparatus. The values of λ, m & a should be given as input. However, the program should ask for three values of "a" and calculate the respective values of the 0's. The output of the program should print, which one of the slits bent the light more. The valid range for the wavelength should be from 380 nm to 750 nm. Out of this range for wavelength should give an output: “Out of the range. Please enter a valid number." 
The wavelength range for different color: 
Wavelength 
Color 
(nm) 
violet 
380-450 
blue 
450-485 
cyan 
485-500 
green 
500-565 
yellow 565-590 
orange 
590-625 
red 
625-750 
Please note the following: 
• λ should be in nm (109) 
d and a's are in μm (10 ̄)